requirements.txt
requests~=2.32.3
apscheduler~=3.11.0
loguru~=0.7.3
beautifulsoup4~=4.13.4
pytz~=2025.2
pydantic-settings~=2.9.1
pydantic~=2.11.4
python-dotenv~=1.1.0
browser-cookie3~=0.20.1
psutil~=7.0.0
uvicorn~=0.34.3
fastapi~=0.115.12